Import of TOMMY/File-Util-3.15 from CPAN.
gitpan-cpan-distribution: File-Util
gitpan-cpan-version:      3.15
gitpan-cpan-path:         TOMMY/File-Util-3.15.tar.gz
gitpan-cpan-author:       TOMMY
gitpan-cpan-maturity:     released

+++ t/6_io.t
@@ -3,79 +3,77 @@ use strict;
 use Test;
 
 # use a BEGIN block so we print our plan before MyModule is loaded
-BEGIN { plan tests => 41, todo => [] }
+BEGIN { plan tests => 12, todo => [] }
 BEGIN { $| = 1 }
 
 # load your module...
 use lib './';
 use File::Util qw( SL NL existent );
 
-my($f)   = File::Util->new();
-my($fh)  = undef;
-my($testbed) = 't/' . $$;
+my($f)    = File::Util->new('--fatals-as-status');
+my($fh)   = undef;
+my($testbed) = 't' . SL . $$;
+my($skip) = !$f->can_write('.') || !$f->can_write('t');
+
+$skip = $skip ? <<'__WHYSKIP__' : $skip;
+Insufficient permissions to perform IO in this directory.  Can't perform tests!
+__WHYSKIP__
 
 # make a temporary testbed directory
-ok($f->make_dir($testbed), $testbed);
+skip($skip, sub { $f->make_dir($testbed) }, $testbed);
 
 # see if it's there
-ok(-e $testbed, 1);
+skip($skip, -e $testbed, 1);
 
 # ...again
-ok($f->existent($testbed), 1);
-
-# directory listing
-{
-   my(@opts) = qw/--files-only --with-paths/;
-   my(@lines) = $f->load_file('MANIFEST','--as-list');
-
-   map { ok($f->existent($_) and -e $_) } $f->list_dir('.', @opts);
-   map {
-      next if !length($_);
-      next if $_ =~ /^META\.yml/o;
-      ok(existent($_) and -e $_)
-   } @lines;
-}
+skip($skip, sub { $f->existent($testbed) }, 1);
 
 # make a temporary file
 my($tmpf) = $testbed . SL . 'tmptst';
-ok($f->write_file('file' => $tmpf, 'content' => $$ . NL),$$ . NL);
+skip(
+	$skip, 
+	sub { 
+		$f->write_file('file' => $tmpf, 'content' => $$ . NL), $$ . NL 
+	}
+);
 
 # get an open file handle
-ok
-   (
-      sub { $fh = $f->open_handle( 'file' => $tmpf, 'mode' => 'append' ); $fh },
-      '/\*File\:\:Util\:\:OPEN\_TO\_FH/'
-   );
+skip( 
+	$skip,
+	sub { $fh = $f->open_handle( 'file' => $tmpf, 'mode' => 'append' ); $fh },
+	'/\*File\:\:Util\:\:OPEN\_TO\_FH/'
+);
 
 # make sure it's still open
-ok(fileno($fh), '/^\d/');
+skip($skip, eval(q{fileno($fh)}), '/^\d/');
 
 # write to it, close it, write to it in append mode
-print( $fh 'Hello world!' . NL ); close($fh);
+unless ($skip) { print( $fh 'Hello world!' . NL ); close($fh); }
 
 # load file
-ok($f->load_file($tmpf),$f->load_file($tmpf));
+skip($skip, sub { $f->load_file($tmpf),$f->load_file($tmpf) });
 
 # write to it with method File::Util::write_file(), compare file contents
 # with the returned value
-ok
-   (
-      $f->write_file
-         (
-            'filename' => $tmpf,
-            'content'  => ( $^O || 'foo' ) . NL,
-            'mode'     => 'append',
-         ),
-   );
+skip (
+	$skip,
+	sub { 
+		$f->write_file(
+			'filename' => $tmpf,
+			'content'  => ( $^O || 'foo' ) . NL,
+			'mode'     => 'append',
+		)
+	},
+);
 
 # get line count of file
-ok($f->line_count($tmpf),3);
+skip($skip, sub { $f->line_count($tmpf) }, 3);
 
 # truncate file
-ok(sub { $f->trunc($tmpf); -s $tmpf }, 0);
+skip($skip, sub { $f->trunc($tmpf); -s $tmpf }, 0);
 
 # get line count of file
-ok($f->line_count($tmpf),0);
+skip($skip, sub { $f->line_count($tmpf)}, 0);
 
 # big directory creation / removal sequence
 my($newdir) =
@@ -86,15 +84,17 @@ my($newdir) =
   . SL . int(rand(time));
 
 # make directories
-ok($f->make_dir($newdir, '--if-not-exists'), $newdir);
+skip($skip, sub { $f->make_dir($newdir, '--if-not-exists') }, $newdir);
 
 # read directories
-my(@items) = $f->list_dir($testbed, '--follow');
+unless ($skip) {
+	my(@items) = $f->list_dir($testbed, '--follow');
 
-# remove directories, temp file, testbed.
-foreach (reverse(sort({ length($a) <=> length($b) } @items)), $testbed) {
+	# remove directories, temp file, testbed.
+	foreach (reverse(sort({ length($a) <=> length($b) } @items)), $testbed) {
 
-   -d $_ ? rmdir($_) || &_rmdie($!) : unlink($_) || &_uldie($!);
+		-d $_ ? rmdir($_) || &_rmdie($!) : unlink($_) || &_uldie($!);
+	}
 }
 
 exit;
